requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
discord.py>=2.0.0 